pyproject.toml
[project]
name = "industrial-agent-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.128.6",
    "grandalf>=0.8",
    "guardrails-ai>=0.8.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "unicorn>=2.1.4",
    "uvicorn>=0.40.0",
]

# Pytest configuration:
# - pythonpath=["."] allows `from app...` imports inside tests
# - testpaths=["tests"] restricts discovery to the tests folder
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
